pass characteristic to handler
To handle multiple hubs, message handlers should be aware of which
device a message was received from.  This change causes the
Characteristic to be passed to Handlers.  The 'path' lens provides
access to the ObjectPath of the Characteristic; this will be unique
per device and it is straightforward to compare to known
Characteristic objects.

A subsequent change to how handlers are registered could allow users
to restrict message delivery to a given device/characteristic.